+++ apps/astarte_data_updater_plant/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix a bug preventing data triggers to be correctly loaded
 
 ## [0.10.0-beta.1] - 2018-08-10
 ### Added